refactor: define `structures` module
