+++ kernel/src/device/pci/xhci/structures/mod.rs
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+pub mod context;
+pub mod dcbaa;
+pub mod registers;
+pub mod ring;
