+++ kernel/src/device/pci/xhci/structures/ring/event/segment_table.rs
